Move ensureDir from shared path utils to CLI file system helper
ensureDir previously lived in src/shared/path-utils.js even though only CLI code invoked it. Relocating it into src/cli/lib/file-system.js keeps the Node-specific directory helper with the CLI and lets shared path utilities avoid pulling in fs dependencies.
